requirements.txt
colorama==0.4.4
numpy==1.21.2
Pillow==8.3.2
requests==2.26.0
scipy==1.7.1